Rename netcdf variable class file
